feat: serve preset passages from server
